+++ next/middleware/middleware.go
@@ -1,14 +1,11 @@
 package middleware
 
 import (
-	"fmt"
-	"github.com/diggerhq/digger/backend/models"
 	"github.com/diggerhq/digger/backend/services"
 	"github.com/gin-gonic/gin"
 	"log"
 	"net/http"
 	"os"
-	"time"
 )
 
 func GetWebMiddleware() gin.HandlerFunc {
@@ -52,30 +49,4 @@ func GetApiMiddleware() gin.HandlerFunc {
 		log.Fatalf("Please specify one of JWT_AUTH or HTTP_BASIC_AUTH")
 		return nil
 	}
-}
-
-func CheckJobToken(c *gin.Context, token string) (*models.JobToken, error) {
-	jobToken, err := models.DB.GetJobToken(token)
-	if jobToken == nil {
-		c.String(http.StatusForbidden, "Invalid bearer token")
-		c.Abort()
-		return nil, fmt.Errorf("invalid bearer token")
-	}
-
-	if time.Now().After(jobToken.Expiry) {
-		log.Printf("Token has already expired: %v", err)
-		c.String(http.StatusForbidden, "Token has expired")
-		c.Abort()
-		return nil, fmt.Errorf("token has expired")
-	}
-
-	if err != nil {
-		log.Printf("Error while fetching token from database: %v", err)
-		c.String(http.StatusInternalServerError, "Error occurred while fetching database")
-		c.Abort()
-		return nil, fmt.Errorf("could not fetch cli token")
-	}
-
-	log.Printf("Token: %v access level: %v", jobToken.Value, jobToken.Type)
-	return jobToken, nil
 }
